Add style decorator
