+++ src/corner_kick/src/index.ts
@@ -5,8 +5,6 @@
 import * as React from 'react';
 import * as ReactDom from 'react-dom';
 
-import './style/base.css';
-
 import { App } from './App';
 
 // Mount the React component App to the div `#react`
