--- src/corner_kick/src/style/base.css
@@ -1,24 +1,0 @@
-/* 
- * Defines the base styling for the application. Imports all other css files in this
- * directory.
- */
-
-@import url('https://fonts.googleapis.com/css?family=Open+Sans');
-
-* {
-    box-sizing: border-box;
-}
-
-body {
-    margin: 0;
-
-    overflow: hidden;
-    user-select: none;
-
-    font-family: 'Open Sans', sans-serif;
-    font-size: 14px;
-}
-
-a {
-    text-decoration: none;
-}
