+++ src/corner_kick/storybook/webpack.config.js
@@ -3,12 +3,14 @@ const path = require('path');
 module.exports = (baseConfig, env, config) => {
     config.module.rules.push({
         test: /\.(ts|tsx)$/,
+        exclude: /node_modules/,
         use: [
             {
                 loader: require.resolve('ts-loader'),
             },
         ],
     });
-    config.resolve.extensions.push('.js', '.ts', '.tsx');
+    config.resolve.extensions.push('.js', '.ts', '.tsx', '.json');
+    config.resolve.alias.SRC = path.resolve(__dirname, '../src');
     return config;
 };
